Split the Cosmoflow model into three files: one for the neural network building blocks, another to stitch together the model, and one to train it.
